adding source freshness tests and adjusting models

+++ models/stg/stg_ordinary__order_items.sql
@@ -3,7 +3,7 @@ with source_data as (
         order_id
         , product_id
         , cast(quantity as int) as quantity
-        , cast(unit_price as float) as unit_price
+        , cast(unit_price as float64) as unit_price
         , cast(created_at as timestamp) as created_at
         , cast(updated_at as timestamp) as updated_at
         , cast(loaded_at as timestamp) as loaded_at
